+++ pr-reports/3482.md
@@ -15,7 +15,10 @@ appearing as generic posts containing only a raw redirect URL.
   the existing X syndication response.
 - Removes an article redirect from post text only when it explicitly maps to
   the parsed article.
-- Renders a native article surface within the established X preview card.
+- Keeps article parsing and rendering in focused modules that preserve the
+  repository's file-size debt ratchet.
+- Renders a native, localized article surface within the established X preview
+  card.
 - Preserves ordinary posts and their `t.co` links when no canonical article
   metadata is present.
